More 64 bit updates for use of MSWindows HANDLE/WHnd representation
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/back/branches/gnustep_testplant_branch@40237 72102866-910b-0410-8b05-ffd578937521

+++ Headers/x11/XGServerWindow.h
@@ -129,7 +129,7 @@ typedef struct _gswindow_device_t {
 @interface XGServer (DPSWindow)
 + (gswindow_device_t *) _windowForXWindow: (Window)xWindow;
 + (gswindow_device_t *) _windowForXParent: (Window)xWindow;
-+ (gswindow_device_t *) _windowWithTag: (int)windowNumber;
++ (gswindow_device_t *) _windowWithTag: (NSInteger)windowNumber;
 - (void) _addExposedRectangle: (XRectangle)rectangle : (NSInteger)win : (BOOL) ignoreBacking;
 - (void) _processExposedRectangles: (NSInteger)win;
 - (void) _initializeCursorForXWindow: (Window) win;
